fix: UsedPoint 타입 변경(BigDecimal → Long)으로 longValue() 호출 제거

+++ src/main/java/one/theone/server/domain/point/service/PointService.java
@@ -67,7 +67,7 @@ public class PointService {
     public void usePoint(Long memberId, Long orderId) {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new ServiceErrorException(OrderExceptionEnum.ERR_ORDER_NOT_FOUND));
-        Long usePoint = order.getUsedPoint().longValue();
+        Long usePoint = order.getUsedPoint();
 
         Long actualBalance = calculateActualBalance(memberId);
         validateBalance(actualBalance, -usePoint);
@@ -98,7 +98,7 @@ public class PointService {
     public void refundPoint(Long memberId, Long orderId) {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new ServiceErrorException(OrderExceptionEnum.ERR_ORDER_NOT_FOUND));
-        Long usedPoint = order.getUsedPoint().longValue();
+        Long usedPoint = order.getUsedPoint();
 
         Long actualBalance = calculateActualBalance(memberId);
